Finish the API integration
When using the AWS console, we get a bunch of stuff for free, but
with terraform we need to add all of the following by hand;

* method reponse with model mapping
* integration response with response template
* permission for API gateway to invoke our lambda function

For the final piece, we need the ID of our AWS account. This can
be set as a variable (like the access key), but since we can
derive it via terraform using aws_caller_identity, we may as well
do that.